Files used to create the files FuzzyParser.java and FuzzyParserSym.java for FuzzyJ
The file FuzzyParser.cup is input for the java_cup program. It produces
2 files (FuzzyParser.java and FuzzyParserSym.java) needed by FuzzyJ to
parse fuzzy expressions. If there are no changes to the types of
expressions used by FuzzyJ then these files will not need to be
regenerated. the file makeFuzzParser.sh shows how to execute the
java_cup program.